DBZ-5222 move files into the API package
