+++ debezium-api/src/main/java/io/debezium/data/Bits.java
