+++ debezium-api/src/main/java/io/debezium/data/VariableScaleDecimal.java
