+++ debezium-api/src/main/java/io/debezium/data/Xml.java
