+++ debezium-api/src/main/java/io/debezium/data/geometry/Geometry.java
