+++ debezium-api/src/main/java/io/debezium/jdbc/DecimalMode.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright Debezium Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.debezium.jdbc;
+
+public enum DecimalMode {
+    PRECISE,
+    DOUBLE,
+    STRING;
+}
