+++ debezium-api/src/test/java/io/debezium/data/EnumSetTest.java
