+++ debezium-api/src/test/java/io/debezium/data/EnumTest.java
