+++ debezium-connector-sqlserver/src/test/java/io/debezium/connector/sqlserver/SqlServerConnectionIT.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 
 import io.debezium.connector.sqlserver.util.TestHelper;
 import io.debezium.doc.FixFor;
-import io.debezium.jdbc.JdbcValueConverters;
+import io.debezium.jdbc.DecimalMode;
 import io.debezium.jdbc.TemporalPrecisionMode;
 import io.debezium.relational.Column;
 import io.debezium.relational.CustomConverterRegistry;
@@ -207,7 +207,7 @@ public class SqlServerConnectionIT {
             Table table = connection.getTableSchemaFromTable(TestHelper.TEST_DATABASE, changeTable);
 
             TableSchemaBuilder tableSchemaBuilder = new TableSchemaBuilder(
-                    new SqlServerValueConverters(JdbcValueConverters.DecimalMode.PRECISE, TemporalPrecisionMode.ADAPTIVE, null),
+                    new SqlServerValueConverters(DecimalMode.PRECISE, TemporalPrecisionMode.ADAPTIVE, null),
                     connection.getDefaultValueConverter(),
                     SchemaNameAdjuster.NO_OP, new CustomConverterRegistry(null), SchemaBuilder.struct().build(), false, false);
 
@@ -377,7 +377,7 @@ public class SqlServerConnectionIT {
             Table table = connection.getTableSchemaFromTable(TestHelper.TEST_DATABASE, changeTable);
 
             TableSchemaBuilder tableSchemaBuilder = new TableSchemaBuilder(
-                    new SqlServerValueConverters(JdbcValueConverters.DecimalMode.PRECISE, TemporalPrecisionMode.ADAPTIVE, null),
+                    new SqlServerValueConverters(DecimalMode.PRECISE, TemporalPrecisionMode.ADAPTIVE, null),
                     connection.getDefaultValueConverter(),
                     SchemaNameAdjuster.NO_OP, new CustomConverterRegistry(null), SchemaBuilder.struct().build(), false, false);
