+++ debezium-core/src/main/java/io/debezium/relational/RelationalDatabaseConnectorConfig.java
@@ -28,8 +28,8 @@ import io.debezium.heartbeat.DatabaseHeartbeatImpl;
 import io.debezium.heartbeat.Heartbeat;
 import io.debezium.heartbeat.HeartbeatConnectionProvider;
 import io.debezium.heartbeat.HeartbeatErrorHandler;
+import io.debezium.jdbc.DecimalMode;
 import io.debezium.jdbc.JdbcConfiguration;
-import io.debezium.jdbc.JdbcValueConverters.DecimalMode;
 import io.debezium.jdbc.TemporalPrecisionMode;
 import io.debezium.relational.Key.CustomKeyMapper;
 import io.debezium.relational.Key.KeyMapper;
